Rename authentication/presentation to views
